GCP implementation
GCP implementation partly done. It has to be as simple as possible, to not add more and more complicated stuff to the already somewhat complex codebase. I think it is very clean now (look at `compare_main.cpp` to get an idea). This is partly done because we need tests to attest the non-trivial new code actually works.