Fixing how the coverages for the correct and incorrect alleles are modelled
Before we were modelling both alleles with the same model. This does not make sense, since in most cases we will have one of the two alleles which are correctly  sequenced, and should be genotyped, and the other due to errors. We now fix this. The coverage for the correct allele is modelled using the model pandora uses for the sample (NegativeBinomial or Binomial). The coverage for the incorrect alleles is modelled using a binomial model with n = mean_kmer_covg and p = error rate. We need to check if this last model should change with the model that pandora uses for a sample, but this is used in Minos, and it seems fine for me... The new KmerCoverageModel class can produce coverages for the correct and the incorrect alleles.